KarnaServer/requirements.txt
Django==4.2.1
django-dbbackup==4.0.2
django-cors-headers==4.0.0
django-env-overrides==0.1.0
django-environ==0.10.0
django-filter==23.2
# django-advanced-filters==2.0.0
django-import-export==3.2.0
# django-guardian==2.4.0
django_mass_edit==3.4.1
django-jsonfield==1.4.1
django-rest-swagger==2.2.0
# django-sundial==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
# django-audit-log==0.7.0
numpy==1.24.3
openpyxl==3.1.2
psycopg2==2.9.6
scipy==1.10.1